+++ lib/jpegli/decode_api_test.cc
@@ -3,7 +3,6 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-#include <setjmp.h>
 #include <stdio.h>
 
 #include <cmath>
