+++ lib/jxl/blending_test.cc
@@ -6,7 +6,6 @@
 #include "lib/extras/codec.h"
 #include "lib/jxl/image_test_utils.h"
 #include "lib/jxl/test_utils.h"
-#include "lib/jxl/testdata.h"
 #include "lib/jxl/testing.h"
 
 namespace jxl {
@@ -16,7 +15,7 @@ using ::testing::SizeIs;
 
 TEST(BlendingTest, Crops) {
   const PaddedBytes compressed =
-      ReadTestData("jxl/blending/cropped_traffic_light.jxl");
+      jxl::test::ReadTestData("jxl/blending/cropped_traffic_light.jxl");
   CodecInOut decoded;
   ASSERT_TRUE(test::DecodeFile({}, Span<const uint8_t>(compressed), &decoded));
   ASSERT_THAT(decoded.frames, SizeIs(4));
@@ -25,7 +24,8 @@ TEST(BlendingTest, Crops) {
   for (const ImageBundle& ib : decoded.frames) {
     std::ostringstream filename;
     filename << "jxl/blending/cropped_traffic_light_frame-" << i << ".png";
-    const PaddedBytes compressed_frame = ReadTestData(filename.str());
+    const PaddedBytes compressed_frame =
+        jxl::test::ReadTestData(filename.str());
     CodecInOut frame;
     ASSERT_TRUE(SetFromBytes(Span<const uint8_t>(compressed_frame), &frame));
     JXL_EXPECT_OK(SamePixels(ib.color(), *frame.Main().color(), _));
